+++ crates/app/src/queries/tasks.rs
@@ -1,11 +1,11 @@
 use super::convert_to_regex;
 use moon_affected::Affected;
-use moon_common::Id;
+use moon_common::{color, Id};
 use moon_task::Task;
 use moon_workspace_graph::WorkspaceGraph;
 use serde::{Deserialize, Serialize};
 use std::{collections::BTreeMap, sync::Arc};
-use tracing::debug;
+use tracing::{debug, warn};
 
 #[derive(Default, Deserialize, Serialize)]
 pub struct QueryTasksOptions {
@@ -42,6 +42,14 @@ fn load_with_regex(
     workspace_graph: &WorkspaceGraph,
     options: &QueryTasksOptions,
 ) -> miette::Result<Vec<Arc<Task>>> {
+    if options.platform.is_some() {
+        warn!(
+            "The {} option is deprecated, use {} instead",
+            color::property("--platform"),
+            color::property("--toolchain"),
+        );
+    }
+
     let id_regex = convert_to_regex("id", &options.id)?;
     let command_regex = convert_to_regex("command", &options.command)?;
     let platform_regex = convert_to_regex("platform", &options.platform)?;